Small logging fix.
git-svn-id: svn://svn.daper.net/moc/trunk@441 910807d9-36e0-0310-a014-e9ea483e2ba4

+++ mp3.c
@@ -192,7 +192,7 @@ static void *mp3_open (const char *file)
 	else {
 		mad_stream_buffer (&data->stream, data->mapped, data->mapped_size);
 		data->stream.error = 0;
-		logit ("mmapped() %ld bytes of file", data->size);
+		logit ("mmapped() %ld bytes of file", (long)data->size);
 	}
 #endif
